feat : Embedded Redis 설정 변경
