+++ chat-service/src/main/java/webrtc/v1/config/EmbeddedRedisConfig.java
@@ -1,0 +1,32 @@
+package webrtc.v1.config;
+
+import java.io.IOException;
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import redis.embedded.RedisServer;
+
+@Profile("local") // profile이 local일때만 활성화
+@Configuration
+public class EmbeddedRedisConfig {
+
+  @Value("${spring.redis.port}")
+  private int redisPort;
+
+  private RedisServer redisServer;
+
+  @PostConstruct
+  public void redisServer() throws IOException {
+    redisServer = new RedisServer(redisPort);
+    redisServer.start();
+  }
+
+  @PreDestroy
+  public void stopRedis() {
+    if (redisServer != null) {
+      redisServer.stop();
+    }
+  }
+}
